Update older versions of Constant
